Extract createTreeVisibilitySlice from projectStore (P5 store slice)
